requirements.txt
pandas
python-dotenv
requests
flask 
flask_sqlalchemy 
flask_cors
Config
pymysql
flask-bcrypt